rename: rename folders and files
