+++ src/components/ProductSectionHeader.js
